+++ app/templates/src/app/core/configs/configs.module.js
@@ -4,7 +4,9 @@
   angular
     .module('<%= prompts.prefix %>.core.configs', [
       '<%= prompts.prefix %>.core.configs.Angular',
-      '<%= prompts.prefix %>.core.configs.ThirdParty'
+      '<%= prompts.prefix %>.core.configs.Translate',
+      '<%= prompts.prefix %>.core.configs.Router',
+      '<%= prompts.prefix %>.core.configs.Promise'
     ]);
 
 })();