+++ app/templates/src/app/core/configs/router.config.js
@@ -1,0 +1,14 @@
+(function () {
+  'use strict';
+
+  angular
+    .module('<%= prompts.prefix %>.core.configs.Router', [])
+    .config(UiRouterConfig);
+
+  function UiRouterConfig($urlRouterProvider){
+    // when there is an empty route, redirect to /index
+    $urlRouterProvider.when('', '/home');
+    $urlRouterProvider.otherwise('/home');
+  }
+
+}());