+++ app/templates/tasks/build.js
@@ -6,4 +6,4 @@ var gulp = require('gulp');
  * BUILD
  */
 
-gulp.task('build', [<% if(!prompts.useTypescript) { %>'jshint', <% } %>'inject', 'build-config']);
+gulp.task('build', [<% if(!prompts.useTypescript) { %>'jshint', <% } %>'bootlint', 'htmllint', 'inject', 'build-config']);