+++ app/templates/tasks/js/jshint.js
@@ -2,7 +2,7 @@
 
 var gulp = require('gulp');
 var projectConfig = require(process.cwd() + '/project.config.js')();
-var $ = require('gulp-load-plugins')({lazy: true});
+var jshint = require('gulp-jshint');
 var stylish = require('jshint-stylish');
 
 /**
@@ -12,7 +12,6 @@ var stylish = require('jshint-stylish');
 gulp.task('jshint', function () {
   return gulp
     .src(projectConfig.path.srcDir + '/' + projectConfig.path.app.scripts)
-    .pipe($.jshint())
-    .pipe($.jshint.reporter(stylish))
-    .pipe($.jshint.reporter('fail'));
+    .pipe(jshint())
+    .pipe(jshint.reporter(stylish));
 });